+++ api/operator/v1beta1/vmsingle_types.go
@@ -494,3 +494,22 @@ func (cr *VMSingle) GetStatusMetadata() *StatusMetadata {
 func (cr *VMSingle) GetAdditionalService() *AdditionalServiceSpec {
 	return cr.Spec.ServiceSpec
 }
+
+// RBACNamespaces returns the namespaces to watch for RBAC resources.
+func (cr *VMSingle) RBACNamespaces(watchedNamespaces []string) []string {
+	if len(watchedNamespaces) == 0 {
+		return nil
+	}
+
+	if cr.Spec.SelectAllByDefault ||
+		cr.Spec.ServiceScrapeNamespaceSelector != nil ||
+		cr.Spec.PodScrapeNamespaceSelector != nil ||
+		cr.Spec.ProbeNamespaceSelector != nil ||
+		cr.Spec.NodeScrapeNamespaceSelector != nil ||
+		cr.Spec.StaticScrapeNamespaceSelector != nil ||
+		cr.Spec.ScrapeConfigNamespaceSelector != nil {
+		return watchedNamespaces
+	}
+
+	return []string{cr.Namespace}
+}